Circles must intersect to get a results
Some examples use abs() to avoid NaN result. This first seems like a
good idea. However if the data is not near perfect it gives a false
assumption that location can be determined from the given data.

Better solution would be to make sure given circle intersect on one
or two points.